% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square a numeric vector}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of the same length as \code{x}, with each element squared.
}
\description{
Squares a single numeric value or a numeric vector
}
\examples{
square(1:5)
square(c(0.5, 2))
}
